pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "copywrite"
version = "0.2.0"
description = "Generate music in the style of reference audio using RAVE and MusicGen"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "librosa>=0.10",
    "soundfile>=0.12",
    "rich>=13.0",
    "torch>=2.0",
    "torchaudio>=2.0",
]

[project.optional-dependencies]
rave = ["acids-rave"]
musicgen = [
    "transformers>=4.30",
    "peft>=0.5",
    "accelerate>=0.20",
    "datasets",
    "audiocraft",
]
dev = ["pytest>=7.0", "pytest-cov"]
all = ["copywrite[rave,musicgen,dev]"]

[project.scripts]
copywrite = "copywrite.cli:cli"

[tool.setuptools.packages.find]
include = ["copywrite*"]
